requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-dotenv==1.0.0
psycopg2_binary==2.9.7
django_extensions==3.2.3
django-filter==23.2
model-mommy==2.0.0
